+++ packages/probes/src/sectum/probes/__init__.py
@@ -18,6 +18,7 @@ from sectum.probes.detection import (
 )
 from sectum.probes.erasure import ErasureProbe, ErasureReport, SurfaceErasure
 from sectum.probes.lora_cross_tenant import LoraCrossTenantProbe
+from sectum.probes.memory_contam import MemoryContamProbe
 from sectum.probes.rag_entity_bleed import RagEntityBleedProbe
 from sectum.probes.semantic_cache import SemanticCacheProbe
 from sectum.probes.tenant_boundary import TenantBoundaryProbe
@@ -33,6 +34,7 @@ __all__ = [
     "Judge",
     "JudgeVerdict",
     "LoraCrossTenantProbe",
+    "MemoryContamProbe",
     "Probe",
     "ProbeRegistry",
     "RagEntityBleedProbe",